Rename .sdk.power_distribution and move actor
Rename frequenz.sdk.power_distribution to frequenz.sdk.power and expose
symbols only in this main module. Also moves the PowerDistributor to
frequenz.sdk.actor.powerdistributing.PowerDistributingActor, including
Request and Result.
