+++ src/frequenz/sdk/power/__init__.py
@@ -1,0 +1,16 @@
+# License: MIT
+# Copyright © 2022 Frequenz Energy-as-a-Service GmbH
+
+"""Utilities to manage power in a microgrid."""
+
+from ._distribution_algorithm import (
+    DistributionAlgorithm,
+    DistributionResult,
+    InvBatPair,
+)
+
+__all__ = [
+    "DistributionAlgorithm",
+    "DistributionResult",
+    "InvBatPair",
+]
