--- src/frequenz/sdk/power_distribution/__init__.py
@@ -1,15 +1,0 @@
-# License: MIT
-# Copyright © 2022 Frequenz Energy-as-a-Service GmbH
-
-"""Distribute power between many batteries.
-
-When charge/discharge method is called the power should be distributed so that
-the SoC in batteries stays at the same level. That way of distribution
-prevents using only one battery, increasing temperature, and maximize the total
-amount power to charge/discharge.
-"""
-
-from .power_distributor import PowerDistributor
-from .utils import Request, Result
-
-__all__ = ["Result", "Request", "PowerDistributor"]
