+++ tests/power/__init__.py
