+++ tests/power/test_distribution_algorithm.py
@@ -20,8 +20,7 @@ from frequenz.api.microgrid.inverter_pb2 import Inverter
 from google.protobuf.timestamp_pb2 import Timestamp  # pylint: disable=no-name-in-module
 
 from frequenz.sdk.microgrid.component_data import BatteryData, InverterData
-from frequenz.sdk.power_distribution.distribution_algorithm import DistributionAlgorithm
-from frequenz.sdk.power_distribution.utils import InvBatPair
+from frequenz.sdk.power import DistributionAlgorithm, InvBatPair
 
 
 @dataclass
